Fix BOTH 32-bit and 64-bit build
Clearly more work is needed to make libriff work with >2 GB files
on 32-bit platforms. However, we couldn't use standard fseek then.

+++ uspace/lib/riff/src/chunk.c
@@ -422,8 +422,8 @@ errno_t riff_rchunk_seek(riff_rchunk_t *rchunk, long offset, int whence)
 		return EINVAL;
 	}
 
-	if (dest < rchunk->ckstart || (unsigned long) dest > rchunk->ckstart +
-	    rchunk->cksize) {
+	if (dest < rchunk->ckstart || (unsigned long) dest >
+	    (unsigned long) rchunk->ckstart + rchunk->cksize) {
 		return ELIMIT;
 	}
